chore: Update sendWithFormData middleware to use upload.none()
This commit updates the sendWithFormData middleware in the message controller to use the upload.none() function from the multer middleware. Previously, the upload.single("file") function was being used, which allowed for file uploads. However, in this case, we don't need to handle file uploads and can use upload.none() instead. This change simplifies the middleware and improves the efficiency of the message sending process.